+++ spb/plot_functions/control.py
@@ -354,9 +354,8 @@ def plot_step_response(
                c: (7, 0, 8),
                d: (6, 0, 25),
                e: (16, 0, 25),
-               # NOTE: remove `None` if using ipywidgets
-               f: (0, 0, 10, 50, None, "lower limit"),
-               g: (10, 0, 25, 50, None, "upper limit"),
+               f: (0, 0, 10, 50, "lower limit"),
+               g: (10, 0, 25, 50, "upper limit"),
            })
 
     See Also
@@ -521,9 +520,8 @@ def plot_impulse_response(
                d: (3, 0, 25),
                e: (12.5, 0, 25),
                f: (17.5, 0, 50),
-               # NOTE: remove `None` if using ipywidgets
-               g: (0, 0, 10, 50, None, "lower limit"),
-               h: (8, 0, 25, 50, None, "upper limit"),
+               g: (0, 0, 10, 50, "lower limit"),
+               h: (8, 0, 25, 50, "upper limit"),
            })
 
 
@@ -688,10 +686,9 @@ def plot_ramp_response(
        params = {
            xi: (6, 0, 10),
            wn: (25, 0, 50),
-           # NOTE: remove `None` if using ipywidgets
-           a: (1, 0, 10, 50, None, "slope"),
-           b: (0, 0, 5, 50, None, "lower limit"),
-           c: (5, 2, 10, 50, None, "upper limit"),
+           a: (1, 0, 10, 50, "slope"),
+           b: (0, 0, 5, 50, "lower limit"),
+           c: (5, 2, 10, 50, "upper limit"),
        }
        plot_ramp_response(
            (tf1, "A"), (tf2, "B"),