+++ tests/interactive/test_panel.py
@@ -127,21 +127,6 @@ def test_DynamicParam():
     r = {a: 1, b: 1, c: True, d: 5, e: 6.1, f: 6}
     assert t.read_parameters() == r
 
-    # raise error because of an invalid formatter. The formatter must be None
-    # or an instance of bokeh.models.formatters.TickFormatter
-    raises(
-        TypeError,
-        lambda: DynamicParam(
-            params={
-                a: (1, 0, 5),
-                b: (2, 1.5, 4.5, 20),
-                c: (3, 2, 5, 30, True, "test1"),
-                d: (1, 1, 10, 10, None, "test2", "log"),
-            },
-            use_latex=True,
-        ),
-    )
-
 
 def test_DynamicParam_symbolic_parameters():
     # verify that we can pass symbolic numbers, which will then be converted